+++ src/plaid/containers/dataset.py
@@ -29,11 +29,12 @@ import numpy as np
 import yaml
 from tqdm import tqdm
 
-from plaid.constants import AUTHORIZED_INFO_KEYS
+from plaid.constants import AUTHORIZED_INFO_KEYS, CGNS_FIELD_LOCATIONS
 from plaid.containers.sample import Sample
 from plaid.containers.utils import check_features_size_homogeneity
 from plaid.types import Array, Feature, FeatureIdentifier
 from plaid.utils.base import DeprecatedError, ShapeError, generate_random_ASCII
+from plaid.utils.deprecation import deprecated
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(
@@ -151,7 +152,7 @@ class Dataset(object):
     # -------------------------------------------------------------------------#
     def get_samples(
         self, ids: Optional[list[int]] = None, as_list: bool = False
-    ) -> dict[int, Sample]:
+    ) -> Union[list[Sample], dict[int, Sample]]:
         """Return dictionnary of samples with ids corresponding to :code:`ids` if specified, else all samples.
 
         Args:
@@ -409,15 +410,19 @@ class Dataset(object):
     def get_field_names(
         self,
         ids: Optional[list[int]] = None,
+        location: Optional[str] = None,
         zone_name: Optional[str] = None,
         base_name: Optional[str] = None,
+        time: Optional[float] = None,
     ) -> list[str]:
         """Return union of fields names in all samples with id in ids.
 
         Args:
             ids (list[int], optional): Select fields depending on sample id. If None, take all samples. Defaults to None.
+            location (str, optional): If provided, only field names from this location will be included. Defaults to None.
             zone_name (str, optional): If provided, only field names from this zone will be included. Defaults to None.
             base_name (str, optional): If provided, only field names containing this base name will be included. Defaults to None.
+            time (float, optional): If provided, only field names from this time will be included. Defaults to None.
 
         Returns:
             list[str]: List of all fields names.
@@ -427,14 +432,30 @@ class Dataset(object):
 
         fields_names = []
         for sample in self.get_samples(ids, as_list=True):
-            times = sample.meshes.get_all_mesh_times()
+            times = [time] if time else sample.meshes.get_all_mesh_times()
             for time in times:
-                f_names = sample.get_field_names(
-                    zone_name=zone_name, base_name=base_name, time=time
+                base_names = (
+                    [base_name]
+                    if base_name
+                    else sample.meshes.get_base_names(time=time)
                 )
-                for f_name in f_names:
-                    if f_name not in fields_names:
-                        fields_names.append(f_name)
+                for base_name in base_names:
+                    zone_names = (
+                        [zone_name]
+                        if zone_name
+                        else sample.meshes.get_zone_names(
+                            time=time, base_name=base_name
+                        )
+                    )
+                    for zone_name in zone_names:
+                        locations = [location] if location else CGNS_FIELD_LOCATIONS
+                        for location in locations:
+                            f_names = sample.get_field_names(
+                                zone_name=zone_name, base_name=base_name, time=time
+                            )
+                            for f_name in f_names:
+                                if f_name not in fields_names:
+                                    fields_names.append(f_name)
         fields_names.sort()
         return fields_names
 
@@ -662,7 +683,7 @@ class Dataset(object):
             )
         return dataset
 
-    def from_features_identifier(
+    def extract_dataset_from_identifier(
         self,
         feature_identifiers: Union[FeatureIdentifier, list[FeatureIdentifier]],
     ) -> Self:
@@ -684,9 +705,25 @@ class Dataset(object):
         dataset.set_infos(copy.deepcopy(self.get_infos()))
 
         for id in self.get_sample_ids():
-            extracted_sample = self[id].from_features_identifier(feature_identifiers)
+            extracted_sample = self[id].extract_sample_from_identifier(
+                feature_identifiers
+            )
             dataset.add_sample(sample=extracted_sample, id=id)
         return dataset
+
+    @deprecated(
+        "Use extract_dataset_from_identifier() instead",
+        version="0.1.8",
+        removal="0.2",
+    )
+    def from_features_identifier(
+        self,
+        feature_identifiers: Union[FeatureIdentifier, list[FeatureIdentifier]],
+    ) -> Self:
+        """DEPRECATED: Use extract_dataset_from_identifier() instead."""
+        return self.extract_dataset_from_identifier(
+            feature_identifiers
+        )  # pragma: no cover
 
     def get_tabular_from_homogeneous_identifiers(
         self,
@@ -744,22 +781,28 @@ class Dataset(object):
 
         return tabular
 
-    def from_tabular(
+    def add_features_from_tabular(
         self,
         tabular: Array,
         feature_identifiers: Union[FeatureIdentifier, list[FeatureIdentifier]],
         restrict_to_features: bool = True,
     ) -> Self:
-        """Generates a dataset from tabular data and feature_identifiers.
+        """Add or update features in the dataset from tabular data using feature identifiers.
+
+        This method takes tabular data and applies it to the dataset, either by updating existing features
+        or adding new ones based on the provided feature identifiers. The method can either:
+        1. Extract only the specified features and return a new dataset with just those features (if restrict_to_features=True)
+        2. Update the specified features in the current dataset while keeping all other existing features (if restrict_to_features=False)
 
         Parameters:
             tabular (Array): of size (nb_sample, nb_features) or (nb_sample, nb_features, dim_feature) if dim_feature>1
-            feature_identifiers (dict or list of dict): One or more feature identifiers.
-            extract_features (bool, optional): If True, only returns the features from feature identifiers, otherwise keep the other features as well
+            feature_identifiers (dict or list of dict): One or more feature identifiers specifying which features to update/add.
+            restrict_to_features (bool, optional): If True, only returns the features from feature identifiers, otherwise keep the other features as well. Defaults to True.
 
         Returns:
-            Self
-                A new dataset defined from tabular data and feature_identifiers.
+            Self: A new dataset with features updated/added from the tabular data. If restrict_to_features=True,
+                  contains only the specified features. If restrict_to_features=False, contains all original
+                  features plus the updated/added ones.
 
         Raises:
             AssertionError
@@ -775,7 +818,7 @@ class Dataset(object):
         features = {id: tabular[i] for i, id in enumerate(self.get_sample_ids())}
 
         if restrict_to_features:
-            dataset = self.from_features_identifier(feature_identifiers)
+            dataset = self.extract_dataset_from_identifier(feature_identifiers)
             dataset.update_features_from_identifier(
                 feature_identifiers=feature_identifiers,
                 features=features,
@@ -789,6 +832,22 @@ class Dataset(object):
             )
 
         return dataset
+
+    @deprecated(
+        "Use add_features_from_tabular() instead",
+        version="0.1.8",
+        removal="0.2",
+    )
+    def from_tabular(
+        self,
+        tabular: Array,
+        feature_identifiers: Union[FeatureIdentifier, list[FeatureIdentifier]],
+        restrict_to_features: bool = True,
+    ) -> Self:
+        """DEPRECATED: Use add_features_from_tabular() instead."""
+        return self.add_features_from_tabular(
+            tabular, feature_identifiers, restrict_to_features
+        )  # pragma: no cover
 
     # -------------------------------------------------------------------------#
     def add_info(self, cat_key: str, info_key: str, info: str) -> None: